pypi package + organize source tree
