+++ .gitignore
@@ -1,2 +1,4 @@
 .env
 __pycache__
+build
+secure_rag.egg-info
