+++ src/rbac_manager.py
@@ -3,9 +3,9 @@ import time
 from bson.objectid import ObjectId
 from dotenv import load_dotenv
 from openai import OpenAI
-from roles import Role
-from utils import embed_documents, get_embedding, get_mongodb_client
-from _types import TextAndRoles, NumpyArray, ToUpload, RetrievedObject
+from src.roles import Role
+from src.utils import embed_documents, get_embedding, get_mongodb_client
+from src._types import TextAndRoles, NumpyArray, ToUpload, RetrievedObject
 
 load_dotenv(override=True)
